requirements.txt
fastapi[standard]==0.115.12
PyJWT==2.10.1
newrelic==10.12.0
python-dotenv==1.1.0
motor==3.7.1
scrapesome